requirements.txt
pip
psutil
net-tools
dnspython

